feat: Add hourmeter value field to fleet vehicle odometer view
This commit adds a new field called "hm_value" to the fleet vehicle odometer view. The "hm_value" field represents the hourmeter value of a vehicle, allowing users to record and track the number of hours a vehicle has been in operation. This feature provides valuable data for maintenance and repair purposes, enabling better monitoring and management of vehicle usage.