Add vehicle approver fields to RepairOrder model
This commit adds two fields `approver_1_id` and `approver_2_id` to the `RepairOrder` model in order to store the information of the approvers for repair orders. These fields are used to track the approvers in the repair report and maintenance report templates. The default values for `approver_1_label` and `approver_2_label` are set as "Kepala Mekanik" (Head Mechanic) and "Operations SPV" (Operations Supervisor), respectively. This change improves the tracking and management of repair orders by including the names of the approvers in the reports.